+++ neuro_san/client/thinking_file_message_processor.py
@@ -28,9 +28,9 @@ from time import strftime
 from time import time
 
 from leaf_common.parsers.dictionary_extractor import DictionaryExtractor
+from neuro_san.internals.journals.origination import Origination
 from neuro_san.message.processors.message_processor import MessageProcessor
 from neuro_san.message.types.chat_message_type import ChatMessageType
-from neuro_san.message.utils.origination import Origination
 
 
 # pylint: disable=too-many-arguments,too-many-positional-arguments
